# contracts/rest/expense-core.openapi.json
{
  "openapi": "3.1.1",
  "info": {"title": "Expense Core API", "version": "1.0.0"},
  "servers": [{"url": "/expense-core/v1"}],
  "paths": {
...
    "/groups/{groupId}": {"x-implementation-status": "implemented-durable", "parameters": [{"$ref": "#/components/parameters/GroupId"}], "get": {"operationId": "getGroup", "responses": {"200": {"description": "Group details", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Group"}}}}}}, "patch": {"operationId": "updateGroup", "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdateGroup"}}}}, "responses": {"200": {"description": "Updated group", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Group"}}}}, "404": {"description": "Group not found or not a member"}}}},
...
    "/groups/{groupId}/archive": {"x-implementation-status": "implemented-durable", "parameters": [{"$ref": "#/components/parameters/GroupId"}], "post": {"operationId": "archiveGroup", "responses": {"200": {"description": "Group archived", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Group"}}}}, "404": {"description": "Group not found or not a member"}, "409": {"$ref": "#/components/responses/Conflict"}}}},
    "/groups/{groupId}/members": {"x-implementation-status": "implemented-durable", "parameters": [{"$ref": "#/components/parameters/GroupId"}], "get": {"operationId": "listGroupMembers", "responses": {"200": {"description": "Group members", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/GroupMember"}}}}}}}},
    "/groups/{groupId}/placeholders": {"x-implementation-status": "implemented-durable", "parameters": [{"$ref": "#/components/parameters/GroupId"}], "post": {"operationId": "createPlaceholder", "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreatePlaceholder"}}}}, "responses": {"201": {"description": "Placeholder created", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/GroupMember"}}}}, "409": {"$ref": "#/components/responses/Conflict"}}}},
    "/groups/{groupId}/members/{membershipId}": {"x-implementation-status": "implemented-durable", "parameters": [{"$ref": "#/components/parameters/GroupId"}, {"name": "membershipId", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}], "delete": {"operationId": "removeGroupMember", "responses": {"204": {"description": "Member removed"}, "404": {"description": "Group or member not found"}, "409": {"$ref": "#/components/responses/Conflict"}}}},
    "/groups/{groupId}/invites": {"parameters": [{"$ref": "#/components/parameters/GroupId"}], "post": {"operationId": "createInvite", "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateInvite"}}}}, "responses": {"201": {"description": "Invite created", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Invite"}}}}}}},
    "/groups/{groupId}/invites/{token}/revoke": {"x-implementation-status": "implemented-durable", "parameters": [{"$ref": "#/components/parameters/GroupId"}, {"$ref": "#/components/parameters/InviteToken"}], "post": {"operationId": "revokeInvite", "responses": {"204": {"description": "Invite revoked"}, "404": {"description": "Group or invite not found"}, "409": {"$ref": "#/components/responses/Conflict"}}}},
    "/invites/{token}/claim": {"parameters": [{"$ref": "#/components/parameters/InviteToken"}], "post": {"operationId": "claimInvite", "responses": {"200": {"description": "Invite claimed", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Group"}}}}, "409": {"$ref": "#/components/responses/Conflict"}}}},
    "/groups/{groupId}/expenses": {"x-implementation-status": "implemented-durable", "parameters": [{"$ref": "#/components/parameters/GroupId"}], "post": {"operationId": "createExpense", "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateExpense"}}}}, "responses": {"201": {"description": "Created", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Expense"}}}}, "409": {"$ref": "#/components/responses/Conflict"}}}, "get": {"operationId": "listExpenses", "parameters": [{"$ref": "#/components/parameters/Cursor"}, {"$ref": "#/components/parameters/Category"}], "responses": {"200": {"description": "Expense page"}}}},
    "/groups/{groupId}/expenses/{expenseId}": {"x-implementation-status": "implemented-durable", "parameters": [{"$ref": "#/components/parameters/GroupId"}, {"name": "expenseId", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}], "put": {"operationId": "updateExpense", "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdateExpense"}}}}, "responses": {"200": {"description": "Updated", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Expense"}}}}, "409": {"$ref": "#/components/responses/Conflict"}}}, "delete": {"operationId": "deleteExpense", "responses": {"204": {"description": "Deleted"}, "409": {"$ref": "#/components/responses/Conflict"}}}},
...
    "/groups/{groupId}/balances": {"x-implementation-status": "implemented-durable", "get": {"operationId": "getBalances", "parameters": [{"$ref": "#/components/parameters/GroupId"}], "responses": {"200": {"description": "Currency-separated balances"}}}},
...
  },
  "components": {"parameters": {"GroupId": {"name": "groupId", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}, "ScheduleId": {"name": "scheduleId", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}, "SettlementId": {"name": "settlementId", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}, "InviteToken": {"name": "token", "in": "path", "required": true, "schema": {"type": "string", "minLength": 64, "maxLength": 64}}, "Cursor": {"name": "cursor", "in": "query", "required": false, "schema": {"type": "string"}}, "Limit": {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 1, "maximum": 100, "default": 50}}, "Category": {"name": "category", "in": "query", "required": false, "schema": {"$ref": "#/components/schemas/ExpenseCategory"}}, "IdempotencyKey": {"name": "Idempotency-Key", "in": "header", "required": true, "schema": {"type": "string", "minLength": 16, "maxLength": 200}}}, "schemas": {
    "Money": {"type": "object", "required": ["currency", "minor"], "properties": {"currency": {"type": "string", "pattern": "^[A-Z]{3}$"}, "minor": {"type": "string", "pattern": "^-?[0-9]+$"}}},
...
    "Group": {"type": "object", "required": ["groupId", "name", "revision"], "properties": {"groupId": {"type": "string", "format": "uuid"}, "name": {"type": "string"}, "revision": {"type": "integer", "minimum": 0}}},
    "UpdateGroup": {"type": "object", "required": ["name"], "properties": {"name": {"type": "string", "minLength": 1, "maxLength": 120}}},
    "GroupMember": {"type": "object", "required": ["membershipId", "groupId", "subject"], "properties": {"membershipId": {"type": "string", "format": "uuid"}, "groupId": {"type": "string", "format": "uuid"}, "subject": {"type": "string"}}},
    "CreateInvite": {"type": "object", "required": ["expiresInHours"], "properties": {"expiresInHours": {"type": "integer", "minimum": 1, "maximum": 168}}},
    "Invite": {"type": "object", "required": ["token", "expiresAt"], "properties": {"token": {"type": "string"}, "expiresAt": {"type": "string", "format": "date-time"}}},
    "CreateExpense": {"type": "object", "required": ["expenseId", "description", "amount", "payers", "allocation"], "properties": {"expenseId": {"type": "string", "format": "uuid"}, "description": {"type": "string", "minLength": 1, "maxLength": 240}, "category": {"$ref": "#/components/schemas/ExpenseCategory", "default": "other"}, "amount": {"$ref": "#/components/schemas/Money"}, "payers": {"type": "array", "minItems": 1, "items": {"type": "object", "required": ["participantId", "amount"], "properties": {"participantId": {"type": "string", "format": "uuid"}, "amount": {"$ref": "#/components/schemas/Money"}}}}, "allocation": {"type": "object", "required": ["mode", "items"], "properties": {"mode": {"type": "string", "enum": ["EQUAL", "EXACT", "PERCENT_BASIS_POINTS", "WEIGHTED_SHARES"]}, "items": {"type": "array", "items": {"type": "object", "required": ["participantId", "value"], "properties": {"participantId": {"type": "string", "format": "uuid"}, "value": {"type": "string", "pattern": "^[0-9]+$"}}}}}}}},
    "UpdateExpense": {"type": "object", "required": ["version", "description", "amount", "payers", "allocation"], "properties": {"version": {"type": "integer", "minimum": 1}, "description": {"type": "string", "minLength": 1, "maxLength": 240}, "category": {"$ref": "#/components/schemas/ExpenseCategory", "default": "other"}, "amount": {"$ref": "#/components/schemas/Money"}, "payers": {"type": "array", "minItems": 1, "items": {"type": "object", "required": ["participantId", "amount"], "properties": {"participantId": {"type": "string", "format": "uuid"}, "amount": {"$ref": "#/components/schemas/Money"}}}}, "allocation": {"type": "object", "required": ["mode", "items"], "properties": {"mode": {"type": "string", "enum": ["EQUAL", "EXACT", "PERCENT_BASIS_POINTS", "WEIGHTED_SHARES"]}, "items": {"type": "array", "items": {"type": "object", "required": ["participantId", "value"], "properties": {"participantId": {"type": "string", "format": "uuid"}, "value": {"type": "string", "pattern": "^[0-9]+$"}}}}}}}},
    "ExpenseCategory": {"type": "string", "enum": ["food", "lodging", "transport", "entertainment", "shopping", "bills", "health", "other"]},
    "Expense": {"type": "object", "required": ["expenseId", "version", "amount", "allocations"], "properties": {"expenseId": {"type": "string", "format": "uuid"}, "version": {"type": "integer", "minimum": 1}, "amount": {"$ref": "#/components/schemas/Money"}, "category": {"$ref": "#/components/schemas/ExpenseCategory"}, "allocations": {"type": "array", "items": {"type": "object"}}}},
...
  }, "responses": {"Conflict": {"description": "Stale version or idempotency conflict"}}}
}